Distribución de proyecto
Se implementó la distribución del proyecto en diferentes archivos separando las funciones para mejorar la legilibilidad y la organización del código.
