+++ GreedGroupX.depend
@@ -1,0 +1,8 @@
+# depslib dependency file v1.0
+1729045876 source:c:\users\emili\documents\programación i\ejercicios\trabajosenclase\greedgroupx\greedgroupx\main.cpp
+	<iostream>
+	<ctime>
+	"funciones.h"
+
+1729045889 c:\users\emili\documents\programación i\ejercicios\trabajosenclase\greedgroupx\greedgroupx\funciones.h
+
