+++ GreedGroupX.layout
@@ -2,9 +2,9 @@
 <CodeBlocks_layout_file>
 	<FileVersion major="1" minor="0" />
 	<ActiveTarget name="Debug" />
-	<File name="main.cpp" open="0" top="0" tabpos="0" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+	<File name="main.cpp" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="-10" zoom_2="0">
 		<Cursor>
-			<Cursor1 position="2042" topLine="0" />
+			<Cursor1 position="4483" topLine="0" />
 		</Cursor>
 	</File>
 </CodeBlocks_layout_file>
