+++ funciones.h
@@ -1,0 +1,9 @@
+# pragma once
+
+void tiradaBloqueadores(int v[], int t);
+void tiradaDeDados(int vD[], int tam);
+int dadosNuevos(int vD[5], int vB[2]);
+void mostrarDados(int vD[], int nT);
+int sumarDados(int vD[], int t);
+
+
